+++ lib/model/counter.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class Counter with ChangeNotifier {
+  int _counter = 0;
+  int get counter => _counter;
+  void incremenet() {
+    _counter++;
+    notifyListeners();
+  }
+}
