+++ pubspec.yaml
@@ -27,12 +27,10 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  
   cupertino_icons: ^1.0.2
-  
   flutter:
     sdk: flutter
-  
+  provider: ^6.0.2
 
 dev_dependencies:
   flutter_lints: ^1.0.0
@@ -48,7 +46,6 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
-
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
@@ -63,10 +60,9 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
   # fonts:
-
-    #- family: Allison
-      #fonts:
-        #- asset: lib/fonts/Allison-Regular.ttf
+  #- family: Allison
+  #fonts:
+  #- asset: lib/fonts/Allison-Regular.ttf
   # fonts:
   #   - family: Schyler
   #     fonts:
