+++ test/widget_test.dart
@@ -6,7 +6,8 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter/material.dart';
-import 'package:flutter_artisan/main.dart';
+
+import 'package:flutter_artisan/view/my_app.dart';
 
 import 'package:flutter_test/flutter_test.dart';
